// src/types/ClipPlayDetails.ts
export interface ClipPlayDetails {
  EventName: string;
  HomeTeamName: string;
  VisitingTeamName: string;
  BatterID: number;
  BatterName: string;
  BatterJerseyNum: number;
  PitcherID: number;
  PitcherName: string;
  PitcherJerseyNum: number;
  TopBottom: string;
  Inning: number;
  Balls: number;
  Strikes: number;
  Outs: number;
  PitchType: string;
  PitchSpeed: number;
  ScoringVisitingTeamID: number;
  ScoringHomeTeamID: number;
  VisitingTeamRuns: number;
  HomeTeamRuns: number;
  ScoringTournamentID: number;
}